Create web_scraping_to_download_amazon_product_images
To ready the training data for a Natural Language Processing model, training data was needed. Amazon doesn't provide the category of the product which are being sold, and manually labeling all the products by going to the amazon site was a very time taking task.
Here ASIN numbers are taken and amazon is searched with those ASINs and then the corresponding images are being downloaded., and stored in an excel.
Now manually labeling the category is much easy as I have the images of the products along with their names.